♻️ [refactor]: 메인화면 루트 경로 변경
