+++ next.config.mjs
@@ -3,8 +3,8 @@ const nextConfig = {
   async redirects() {
     return [
       {
-        source: '/',
-        destination: '/main',
+        source: '/main',
+        destination: '/',
         permanent: true,
       },
     ]
